Move defaultRestoreHandler into client
